backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
scikit-learn
pdfplumber
PyPDF2
requests
python-multipart